backend/requirements.txt
# Backend specific requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
twilio==8.11.1
httpx==0.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.36.0
openai==2.8.1
websockets==12.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
redis==5.0.1
# Bulk Campaigns
phonenumbers==8.13.27
pandas==2.1.4
openpyxl==3.1.2
APScheduler==3.10.4
pypdf==5.3.0
python-docx==1.1.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0